+++ packages/tenbo/test/index.js
@@ -1,6 +1,6 @@
 import test from '../node_modules/tehanu/dist/suite.min.mjs?name=sum'
-import 'tehanu-repo-tape/dist/index.min'
-import { equal } from 'tehanu-teas/dist/index.min'
+import '../node_modules/tehanu-repo-tape/dist/index.min.mjs'
+import { equal } from '../node_modules/tehanu-teas/dist/index.min.mjs'
 import sum from './sum.mjs'
 
 test('one number', () => equal(sum(1), 1))